+++ src/client/__init__.py
@@ -1,1 +1,7 @@
-from .opnsense_client import OPNsenseClient, OPNsenseHost  # noqa: F401
+from .opnsense_client import (  # noqa: F401
+    OPNsenseAuthError,
+    OPNsenseClient,
+    OPNsenseError,
+    OPNsenseHost,
+    OPNsenseTimeoutError,
+)